+++ src/categories/models/index.ts
@@ -1,2 +1,3 @@
 export { CategoryListResult } from './category-list-result';
 export { Category } from './category';
+export { CategoryProduct } from './category-product';