+++ src/products/products-client.ts
@@ -1,7 +1,7 @@
 import * as util from 'util';
 import { RestClient } from '../rest-client';
 import { plainToClass } from 'class-transformer';
-import { ListResult, SearchCriteria } from '../common-models';
+import { ListResult, SearchCriteria, SearchFilterGroup } from '../common-models';
 import { ClientBase } from '../client-base';
 import { Product } from './models';
 
@@ -9,6 +9,34 @@ export class ProductsClient extends ClientBase {
 
   constructor(restClient: RestClient) {
     super(restClient);
+  }
+
+  /**
+   * List all products of a specified category.
+   * @method listByCategory
+   * @param  categoryId     The category ID of the products.
+   * @param  pageSize       The number of results to be fetched in one "page" of API response. Default is 9999.
+   * @param  currentPage    The current page of the "pages" of result. Use in conjunction with pageSize.
+   * @return                A list of products.
+   */
+  async listByCategory(
+    categoryId: string | number,
+    pageSize: number = 9999,
+    currentPage: number = 1
+  ): Promise<ListResult<Product>> {
+    const searchCriteria = new SearchCriteria();
+    const filterGroup = new SearchFilterGroup();
+    filterGroup.addSearchFilter('category_id', categoryId, 'eq');
+    searchCriteria.addSearchFilterGroup(filterGroup);
+    const query = `searchCriteria[pageSize]=${pageSize}&searchCriteria[currentPage]=${currentPage}`
+      + `&${searchCriteria.toString()}`;
+    const endpointUrl = util.format('/products?%s', query);
+    try {
+      const result = await this.restClient.get(endpointUrl);
+      return new ListResult<Product>(result, plainToClass(Product, result.items as Product[]));
+    } catch (error) {
+      throw error;
+    }
   }
 
   /**